feature/rewrite-decorator provide tests
