+++ tests/attachStore.spec.js
@@ -1,0 +1,51 @@
+import {AttachStore} from '../src/attachStore';
+import { actions, store, initialState } from './initials';
+
+describe('attachStore', () => {
+  test('TestComponent.testMethod should be called', () => {
+    class TestComponent {
+      testMethod() {}
+    }
+
+    const component = new TestComponent();
+
+    const observed = state => ({
+      price: state.price,
+    });
+
+    const passive = state => ({
+      stockInfo: state.stockInfo,
+    });
+
+    const spy = jest.spyOn(TestComponent.prototype, 'testMethod');
+
+    TestComponent.prototype.subscriptionQueue = [{
+      observed,
+      passive,
+      method: TestComponent.prototype.testMethod,
+      }];
+
+    AttachStore(store)(TestComponent);
+    component.init();
+
+    store.dispatch({ type: actions.observed });
+
+    expect(spy).toHaveBeenCalledWith({
+      price: 120,
+      stockInfo: 'InStock',
+    });
+
+    store.dispatch({ type: actions.observed });
+    expect(spy).toHaveBeenCalledTimes(2);
+
+    store.dispatch({ type: actions.static });
+    expect(spy).toHaveBeenCalledTimes(2);
+
+    store.dispatch({ type: actions.observed });
+    expect(spy).toHaveBeenCalledTimes(3);
+    expect(spy).toHaveBeenCalledWith({
+      price: 160,
+      stockInfo: 'OutOfStock',
+    });
+  });
+});
